refactor: extract run-types.ts and run-utils.ts from run.ts (#447)
Step 1-2 of run.ts decomposition:
- run-types.ts: RunOptions, ExecutionContext interfaces + constants
- run-utils.ts: pure utility functions (generateExecutionId, selectMcpConfig,
  detectTaskType, resolveModel, ensureProjectTrusted, getProjectRoot,
  formatDuration, checkClaudeCliAvailable, getClaudeModelAlias)
